+++ runtime/browser/devtools/xwalk_devtools_delegate.h
@@ -13,33 +13,17 @@
 #include "base/compiler_specific.h"
 #include "base/memory/ref_counted_memory.h"
 #include "base/memory/weak_ptr.h"
-#include "content/public/browser/devtools_http_handler_delegate.h"
 #include "content/public/browser/devtools_manager_delegate.h"
 #include "url/gurl.h"
 #include "xwalk/runtime/browser/runtime.h"
 
-namespace content {
+namespace devtools_http_handler {
 class DevToolsHttpHandler;
 }
 
 namespace xwalk {
 
 class XWalkBrowserContext;
-
-class XWalkDevToolsHttpHandlerDelegate :
-    public content::DevToolsHttpHandlerDelegate {
- public:
-  XWalkDevToolsHttpHandlerDelegate();
-  ~XWalkDevToolsHttpHandlerDelegate() override;
-
-  // DevToolsHttpHandlerDelegate implementation.
-  std::string GetDiscoveryPageHTML() override;
-  bool BundlesFrontendResources() override;
-  base::FilePath GetDebugFrontendDir() override;
-
- private:
-  DISALLOW_COPY_AND_ASSIGN(XWalkDevToolsHttpHandlerDelegate);
-};
 
 class XWalkDevToolsDelegate : public content::DevToolsManagerDelegate,
                               public Runtime::Observer {
@@ -56,22 +40,14 @@ class XWalkDevToolsDelegate : public content::DevToolsManagerDelegate,
   base::DictionaryValue* HandleCommand(
       content::DevToolsAgentHost* agent_host,
       base::DictionaryValue* command_dict) override;
-  scoped_ptr<content::DevToolsTarget> CreateNewTarget(
-      const GURL& url) override;
-  void EnumerateTargets(TargetCallback callback) override;
-  std::string GetPageThumbnailData(const GURL& url) override;
-  void ProcessAndSaveThumbnail(const GURL& url,
-                               scoped_refptr<base::RefCountedBytes> png);
 
  private:
   // Runtime::Observer
   void OnNewRuntimeAdded(Runtime* runtime) override;
   void OnRuntimeClosed(Runtime* runtime) override;
 
-  using ThumbnailMap = std::map<GURL, std::string>;
-  ThumbnailMap thumbnail_map_;
   XWalkBrowserContext* browser_context_;
-  base::WeakPtrFactory<XWalkDevToolsDelegate> weak_factory_;
+
   DISALLOW_COPY_AND_ASSIGN(XWalkDevToolsDelegate);
 };
