+++ xwalk.gyp
@@ -38,6 +38,7 @@
         '../base/third_party/dynamic_annotations/dynamic_annotations.gyp:dynamic_annotations',
         '../cc/cc.gyp:cc',
         '../components/components.gyp:autofill_content_renderer',
+        '../components/components.gyp:devtools_http_handler',
         '../components/components.gyp:visitedlink_browser',
         '../components/components.gyp:visitedlink_renderer',
         '../content/content.gyp:content',
